Move lego.app.objectpermissions to lego.permissions and move the model permissions from users there
