+++ lego/permissions/tests/model.py
@@ -2,7 +2,7 @@
 from django.db import models
 from django.utils.translation import ugettext_lazy as _
 
-from lego.app.objectpermissions.models import ObjectPermissionsModel
+from lego.permissions.models import ObjectPermissionsModel
 
 
 class TestModel(ObjectPermissionsModel):
