+++ lego/permissions/tests/serializer.py
@@ -1,7 +1,7 @@
 # -*- coding: utf8 -*-
 from rest_framework import serializers
 
-from lego.app.objectpermissions.tests.model import TestModel
+from lego.permissions.tests.model import TestModel
 
 
 class TestSerializer(serializers.ModelSerializer):
